requirements.txt
# Tools
ipython>=5.0.0
numpy>=1.11.2
scipy>=0.18.1
astropy>=1.2.1
pandas>=0.18.1
SQLAlchemy>=1.1.3
SQLAlchemy-Utils>=0.32.9
networkx>=1.11
matplotlib>=1.5.3
Pillow>=3.3.1
requests>=2.10.0
requests-futures>=0.9.7
dogpile.cache>=0.6.2
python-memcached>=1.58
WTForms>=2.1
WTForms-Components>=0.10.0
validators>=0.10.3
intervals>=0.8.0
coveralls>=1.1
sphinx_bootstrap_theme>=0.4.12
sphinxcontrib-httpdomain>=1.5.0
webargs>=1.5.2
fuzzywuzzy>=0.15.0
python-Levenshtein>=0.12.0
yamlordereddictloader>=0.2.2

# Web
blinker>=1.4
Flask>=0.11.1
Flask-Classy>=0.6.10
Flask-JSGlue>=0.3
Flask-FeatureFlags>=0.6
flask-profiler>=1.0.1
Flask-RESTful>=0.3.5
Flask-SQLAlchemy>=2.1
Flask-Testing>=0.6.1
raven>=5.32.0
pgpasslib>=1.1.0
# psycopg2>=2.6.2
PyYAML>=3.11
# uwsgi>=2.0.15
Flask-Limiter>=0.9.4
